Allow playback if resource manager is not available.
If mediaplayer is used in environment which doesn't have
running policy resource manager, mediaplayer is effectively
useless since it never is granted resources and thus
mediaplayer never starts real playback. This patch changes
functionality so that if resource policy manager is not
available in D-Bus bus, all resources are unconditionally
granted.